Move InsightView code generator outside internals
JIRA: RAIL-4195
