+++ libs/sdk-ui-ext/src/insightView/embedding/insightViewCodeGenerator.ts
@@ -1,10 +1,15 @@
 // (C) 2022 GoodData Corporation
 
 import { IInsight, insightRef, insightTitle } from "@gooddata/sdk-model";
-import { IInsightViewProps } from "../../insightView";
-import { getReactEmbeddingCodeGenerator } from "../utils/embeddingCodeGenerator";
+import { getReactEmbeddingCodeGenerator } from "../../internal/utils/embeddingCodeGenerator";
+import { IInsightViewProps } from "../types";
 import { insightViewAdditionalTransformations } from "./insightViewCodeGenUtils";
 
+/**
+ * DO NOT USE THIS INSIGHTVIEW CODE GENERATOR, IT'S FOR INTERNAL PURPOSE ONLY.
+ *
+ * @internal
+ */
 export const insightViewCodeGenerator = getReactEmbeddingCodeGenerator<IInsightViewProps>({
     component: {
         importType: "named",
