+++ libs/sdk-ui-ext/src/insightView/embedding/tests/insightViewCodeGenerator.test.ts
